checkout refactored and made to scale better
